AllPersons.jsx -- PersonCard.jsx -- ArrowBack.jsx -- SearchbarInput.jsx
AllPersons.jsx is the component responsible for rendering all existing people in the database

PersonCard.jsx is the component that will be redered with some person information such as the photo and name. In addiction it will also have the option to edit or delete the person.

ArrowBack.jsx is the component that will be rendered throughout the application, dynamically where it will allow the user to retreat to some point in the application

SearchbarInput.jsx is the component rendered only within AllPerson.jsx that is responsible for receiving some text entered by the user to serve as a search filter for the list of people.