InputImage.jsx created
InputImage.jsx is the component resposible to render the input to insert an image to create or edit a person.

The component detect if it has been given to him a imageName and it allow us to change the photo. Case it won't he display the input to insert a new one.